Weather app 3

+++ wather_app3/lib/model/weather_model.dart
@@ -2,7 +2,7 @@ class Weather {
   String? cityName;
   double? temp;
   double? wind;
-  int? humidity;
+  double? humidity;
   double? feels_like;
   int? pressure;
